requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
python-decouple==3.8
pillow==11.0.0